refactor(sidepanel): extract data transforms module (Phase 1/3)
