+++ pages/sidepanel/sidepanel.js
@@ -9,7 +9,7 @@
 
 /* global chrome */
 
-import { normalizeUrl, computeStableUrl, isRootUrl } from '../../scripts/utils/urlUtils.js';
+import { normalizeUrl, computeStableUrl } from '../../scripts/utils/urlUtils.js';
 import {
   SAVED_PREFIX,
   HIGHLIGHTS_PREFIX,
@@ -30,7 +30,15 @@ import {
   pickAliasCandidate,
   pickHighlightsFromStorage,
 } from '../../scripts/highlighter/core/HighlightLookupResolver.js';
-import { compareKeysAlphabetically } from '../../scripts/utils/keyOrdering.js';
+import {
+  normalizeStorageSnapshot,
+  getHighlightList,
+  buildPageEntry,
+  buildLegacyPageEntry,
+  _buildAliasMap,
+  _collectDeletionKeys,
+  resolveUnsyncedOwnership,
+} from './sidepanel-data-transforms.js';
 
 // === 共享狀態（保留於入口，UI 模組不直接存取） ===
 
@@ -256,389 +264,6 @@ function requireElement(element, elementName) {
 }
 
 // === 業務邏輯 ===
-
-/**
- * @param {*} snapshot
- * @returns {Record<string, any>}
- */
-function normalizeStorageSnapshot(snapshot) {
-  if (!snapshot) {
-    return {};
-  }
-  if (typeof snapshot !== 'object') {
-    return {};
-  }
-  if (Array.isArray(snapshot)) {
-    return {};
-  }
-  return snapshot;
-}
-
-/**
- * @param {*} value
- * @returns {Array}
- */
-function getHighlightList(value) {
-  if (Array.isArray(value)) {
-    return value;
-  }
-  const highlights = value?.highlights;
-  return Array.isArray(highlights) ? highlights : [];
-}
-
-/**
- * @param {Array} highlights
- * @returns {object}
- */
-function buildHighlightSummary(highlights) {
-  return {
-    highlightCount: highlights.length,
-    previewHighlights: UI.buildPreviewHighlights(highlights),
-    remainingCount: Math.max(0, highlights.length - UI.PREVIEW_HIGHLIGHT_COUNT),
-  };
-}
-
-/**
- * @param {object} value
- * @returns {boolean}
- */
-function hasSyncedPageValue(value) {
-  return Boolean(value?.notion?.pageId);
-}
-
-/**
- * @param {object} value
- * @param {string} url
- * @returns {string}
- */
-function resolvePageEntryTitle(value, url) {
-  const notion = value?.notion;
-  if (notion?.title) {
-    return notion.title;
-  }
-  const metadata = value?.metadata;
-  if (metadata?.title) {
-    return metadata.title;
-  }
-  return UI.extractDomain(url);
-}
-
-/**
- * @param {object} value
- * @returns {number}
- */
-function resolvePageLastUpdated(value) {
-  return value?.metadata?.lastUpdated || 0;
-}
-
-/**
- * 從 page_* 對象建立頁面條目（新格式）
- *
- * @param {string} key - storage key
- * @param {string} url - normalized url
- * @param {object} value - page_* 物件
- * @returns {object|null} 頁面條目，若應跳過則返回 null
- */
-function buildPageEntry(key, url, value) {
-  if (hasSyncedPageValue(value)) {
-    return null; // 已同步，跳過
-  }
-  if (isRootUrl(url)) {
-    return null; // 已同步或根路徑，跳過
-  }
-  const highlights = getHighlightList(value);
-  if (highlights.length === 0) {
-    return null; // 無標註不顯示
-  }
-  return {
-    url,
-    storageKey: key,
-    title: resolvePageEntryTitle(value, url),
-    lastUpdated: resolvePageLastUpdated(value),
-    ...buildHighlightSummary(highlights),
-  };
-}
-
-/**
- * @param {object} all
- * @param {string} url
- * @param {string} canonicalUrl
- * @returns {{isSaved: boolean, savedData: any}}
- */
-function resolveLegacySavedState(all, url, canonicalUrl) {
-  const storageSnapshot = normalizeStorageSnapshot(all);
-  const savedDataOriginal = storageSnapshot[`${SAVED_PREFIX}${url}`];
-  if (savedDataOriginal?.notionPageId) {
-    return { isSaved: true, savedData: savedDataOriginal };
-  }
-
-  const savedDataCanonical = storageSnapshot[`${SAVED_PREFIX}${canonicalUrl}`];
-  if (savedDataCanonical?.notionPageId) {
-    return { isSaved: true, savedData: savedDataCanonical };
-  }
-
-  return { isSaved: false, savedData: savedDataOriginal || savedDataCanonical };
-}
-
-/**
- * @param {*} savedData
- * @param {*} value
- * @param {string} url
- * @returns {string}
- */
-function resolveLegacyEntryTitle(savedData, value, url) {
-  if (savedData?.title) {
-    return savedData.title;
-  }
-  if (value?.title) {
-    return value.title;
-  }
-  return UI.extractDomain(url);
-}
-
-/**
- * @param {*} value
- * @returns {number}
- */
-function resolveLegacyLastUpdated(value) {
-  return value?.updatedAt || 0;
-}
-
-/**
- * 從 highlights_* 對象建立頁面條目（舊格式）
- *
- * @param {string} key - storage key
- * @param {string} url - normalized url
- * @param {*} value - highlights_* 值
- * @param {{all: object, aliasMap: Map}} context
- * @returns {object|null}
- */
-function buildLegacyPageEntry(key, url, value, context) {
-  if (isRootUrl(url)) {
-    return null;
-  }
-  const { all, aliasMap } = context;
-  const canonicalUrl = aliasMap.get(url) || url;
-  const { isSaved, savedData } = resolveLegacySavedState(all, url, canonicalUrl);
-  if (isSaved) {
-    return null; // 已同步
-  }
-  const highlights = getHighlightList(value);
-  if (highlights.length === 0) {
-    return null;
-  }
-  return {
-    url,
-    storageKey: key,
-    title: resolveLegacyEntryTitle(savedData, value, url),
-    lastUpdated: resolveLegacyLastUpdated(value),
-    ...buildHighlightSummary(highlights),
-  };
-}
-
-/**
- * 為 unsynced enumeration 解析 canonical ownership。
- *
- * Phase 4：sidepanel 的 unsynced 枚舉不能僅以 raw URL 當作識別維度，
- * 否則 `page_<stableUrl>` 與 `page_<originalUrl>` 會被當成兩張卡片顯示。
- *
- * 此 helper 將 storage 中所有 highlight 相關 key 歸併到單一 canonical identity：
- * - canonical URL = url_alias:<url> 解析結果（與 HighlightLookupResolver 採同一規則）
- * - 對每一組同 canonical 的 keys，依優先順序選出 owner：
- *     1. page_<canonical>（alias 命中時最可靠的 canonical key）
- *     2. 任意其他 page_*（alias 指向空 stable key 時的回退）
- *     3. 任意 highlights_*（舊格式回退）
- *
- * @param {Record<string, any>} allStorageData - chrome.storage.local.get(null) 的結果
- * @returns {Map<string, {ownerKey: string, ownerUrl: string, ownerValue: any, format: 'page'|'legacy'}>}
- *          canonicalUrl → owner descriptor
- */
-/**
- * 從 allStorageData 建立 url → canonicalUrl 的 alias map。
- *
- * @param {Record<string, any>} allStorageData
- * @returns {Map<string, string>}
- * @private
- */
-function _buildAliasMap(allStorageData) {
-  const storageSnapshot = normalizeStorageSnapshot(allStorageData);
-  const aliasMap = new Map();
-  for (const [key, value] of Object.entries(storageSnapshot)) {
-    if (key.startsWith(URL_ALIAS_PREFIX)) {
-      aliasMap.set(key.slice(URL_ALIAS_PREFIX.length), value);
-    }
-  }
-  return aliasMap;
-}
-
-/**
- * Pure helper：根據 contract + reverse alias scan 計算同 canonical group 應刪除的 keys。
- *
- * Phase 4 follow-up（2026-05-03 plan §2、§3）：
- * - Forward alias：用 pickAliasCandidate + resolveHighlightLookupKeys 取得 contract（含 mutationTargetKey、legacyCleanupKeys）
- * - Reverse alias：掃 aliasMap，把所有 value === canonicalUrl 的 entry 對應的
- *   page_<other> / highlights_<other> 一併納入候選集
- * - 最後以 snapshot 過濾，僅保留實際存在的 keys，並沿用 planDeleteCleanup 的字典序輸出邏輯
- *
- * 此 helper 純函數,無 chrome.storage IO；deleteUnsyncedPage / deleteAllUnsyncedPages 共用同一條路徑。
- *
- * @param {string} pageUrl - 由 storage key 抽出的 raw URL（owner.ownerUrl 或 _extractUrlFromStorageKey 結果）
- * @param {string} fallbackKey - 呼叫端傳入的 owner storageKey,即使 contract 沒涵蓋也會嘗試刪除
- * @param {Record<string, any>} allStorageData - chrome.storage.local.get(null) 結果
- * @param {Map<string, string>} aliasMap - URL_ALIAS_PREFIX 解析結果（forward 與 reverse 都用得到）
- * @returns {string[]} 應呼叫 chrome.storage.local.remove 的 key 清單（已字典序）
- */
-function _collectDeletionKeys(pageUrl, fallbackKey, allStorageData, aliasMap) {
-  const storageSnapshot = normalizeStorageSnapshot(allStorageData);
-  // Forward alias：用 pickAliasCandidate 走相同的 alias validation 規則
-  const synthAliasData = {
-    [`${URL_ALIAS_PREFIX}${pageUrl}`]: aliasMap.get(pageUrl) ?? null,
-  };
-  const aliasCandidate = pickAliasCandidate(synthAliasData, pageUrl);
-  const contract = resolveHighlightLookupKeys(pageUrl, aliasCandidate);
-
-  // Reverse alias：找 aliasMap 中所有 value === contract.canonicalUrl 的 from-URL,
-  // 收集對應的 page_<other> / highlights_<other>（不在 contract 內,但屬同 canonical group）
-  const reverseMembers = [];
-  for (const [aliasFromUrl, aliasToUrl] of aliasMap) {
-    if (typeof aliasToUrl !== 'string' || aliasToUrl !== contract.canonicalUrl) {
-      continue;
-    }
-    if (aliasFromUrl === contract.canonicalUrl || aliasFromUrl === pageUrl) {
-      // 已涵蓋於 contract（forward path 與 normalizedUrl）
-      continue;
-    }
-    reverseMembers.push(`${PAGE_PREFIX}${aliasFromUrl}`, `${HIGHLIGHTS_PREFIX}${aliasFromUrl}`);
-  }
-
-  // 候選集 = contract.mutationTargetKey + contract.legacyCleanupKeys + reverseMembers + fallbackKey
-  const candidateSet = new Set([
-    contract.mutationTargetKey,
-    ...contract.legacyCleanupKeys,
-    ...reverseMembers,
-    fallbackKey,
-  ]);
-
-  // 僅保留 snapshot 中實際存在的 key（避免無謂 remove）
-  const result = [...candidateSet].filter(
-    k => typeof k === 'string' && storageSnapshot[k] !== undefined && storageSnapshot[k] !== null
-  );
-
-  return result.toSorted(compareKeysAlphabetically);
-}
-
-/**
- * @param {string} key
- * @param {*} value
- * @returns {{key:string,url:string,value:any,format:'page'|'legacy'} | null}
- * @private
- */
-function _parseHighlightStorageEntry(key, value) {
-  if (key.startsWith(PAGE_PREFIX)) {
-    return { key, url: key.slice(PAGE_PREFIX.length), value, format: 'page' };
-  }
-  if (key.startsWith(HIGHLIGHTS_PREFIX)) {
-    return { key, url: key.slice(HIGHLIGHTS_PREFIX.length), value, format: 'legacy' };
-  }
-  return null;
-}
-
-/**
- * @param {Map<string, {pages: Array<{key:string,url:string,value:any}>, legacies: Array<{key:string,url:string,value:any}>}>} groups
- * @param {string} canonical
- * @returns {{pages: Array<{key:string,url:string,value:any}>, legacies: Array<{key:string,url:string,value:any}>}}
- * @private
- */
-function _ensureStorageEntryGroup(groups, canonical) {
-  if (!groups.has(canonical)) {
-    groups.set(canonical, { pages: [], legacies: [] });
-  }
-  return groups.get(canonical);
-}
-
-/**
- * 將 allStorageData 中的 page_* / highlights_* 按 canonical URL 分群。
- *
- * @param {Record<string, any>} allStorageData
- * @param {Map<string, string>} aliasMap
- * @returns {Map<string, {pages: Array<{key:string,url:string,value:any}>, legacies: Array<{key:string,url:string,value:any}>}>}
- * @private
- */
-function _groupStorageEntries(allStorageData, aliasMap) {
-  const storageSnapshot = normalizeStorageSnapshot(allStorageData);
-  /** @type {Map<string, {pages: Array<{key:string,url:string,value:any}>, legacies: Array<{key:string,url:string,value:any}>}>} */
-  const groups = new Map();
-  for (const [key, value] of Object.entries(storageSnapshot)) {
-    const entry = _parseHighlightStorageEntry(key, value);
-    if (!entry) {
-      continue;
-    }
-
-    const canonical = aliasMap.get(entry.url) || entry.url;
-    const group = _ensureStorageEntryGroup(groups, canonical);
-    if (entry.format === 'page') {
-      group.pages.push({ key: entry.key, url: entry.url, value: entry.value });
-    } else {
-      group.legacies.push({ key: entry.key, url: entry.url, value: entry.value });
-    }
-  }
-  return groups;
-}
-
-/**
- * 從同一 canonical URL 的 group 中選出最優先的 owner descriptor。
- *
- * @param {string} canonicalUrl
- * @param {{pages: Array<{key:string,url:string,value:any}>, legacies: Array<{key:string,url:string,value:any}>}} group
- * @returns {{ownerKey: string, ownerUrl: string, ownerValue: any, format: 'page'|'legacy'} | null}
- * @private
- */
-function _pickGroupOwner(canonicalUrl, group) {
-  // Priority 1: page_<canonical>
-  let chosen = group.pages.find(page => page.url === canonicalUrl) || null;
-
-  // Priority 2: 任意其他 page_*（例如 alias 指向空 stable key 時的 page_<original>）
-  if (!chosen && group.pages.length > 0) {
-    chosen = group.pages[0];
-  }
-
-  if (chosen) {
-    return { ownerKey: chosen.key, ownerUrl: chosen.url, ownerValue: chosen.value, format: 'page' };
-  }
-
-  // Priority 3: highlights_*（舊格式）
-  if (group.legacies.length > 0) {
-    const legacy = group.legacies[0];
-    return {
-      ownerKey: legacy.key,
-      ownerUrl: legacy.url,
-      ownerValue: legacy.value,
-      format: 'legacy',
-    };
-  }
-
-  return null;
-}
-
-function resolveUnsyncedOwnership(allStorageData) {
-  const storageSnapshot = normalizeStorageSnapshot(allStorageData);
-  if (Object.keys(storageSnapshot).length === 0) {
-    return new Map();
-  }
-
-  const aliasMap = _buildAliasMap(storageSnapshot);
-  const groups = _groupStorageEntries(storageSnapshot, aliasMap);
-
-  const owners = new Map();
-  for (const [canonicalUrl, group] of groups) {
-    const owner = _pickGroupOwner(canonicalUrl, group);
-    if (owner) {
-      owners.set(canonicalUrl, owner);
-    }
-  }
-
-  return owners;
-}
 
 /**
  * 從 chrome.storage.local 取出所有未同步頁面的標註資料。
